frontend/requirements.txt
gradio==5.43.1
requests==2.32.5
python-dotenv==1.1.1